service: Test Service.at for hex strings
